back-end/requirements.txt
Flask==3.0.0
pymongo==4.6.0
python-dotenv==1.0.0
marshmallow==3.20.1
marshmallow-mongoengine==0.4.0
flask-cors==4.0.0
python-dateutil==2.8.2
pydantic==2.5.0
pydantic-settings==2.1.0


